+++ Benchmarks/src/Collections/List/ArrayListBenchmarks.cs
@@ -7,8 +7,9 @@ namespace Benchmarks.Collections.List;
 [SuppressMessage("ReSharper", "UnusedMember.Global")]
 [SuppressMessage("ReSharper", "UnusedType.Global")]
 public class ArrayListBenchmarks {
-	public static int Iterations;
-	public static int LoopIterations;
+	public static ulong Iterations;
+	public static ulong LoopIterations;
+
 
 	public static readonly ArrayList Data = new(1000);
 
@@ -21,7 +22,7 @@ public class ArrayListBenchmarks {
 	[Benchmark("ListGet", "Tests getting values sequentially from a ArrayList")]
 	public static int ArrayListGet() {
 		int sum = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i = 0; i < LoopIterations; i++) {
 			for (int j = 0; j < Data.Count; j++) {
 				sum += (int)Data[CollectionsHelpers.SequentialIndices[j]];
 			}
@@ -33,7 +34,7 @@ public class ArrayListBenchmarks {
 	[Benchmark("ListGet", "Tests getting values randomly from a ArrayList")]
 	public static int ArrayListGetRandom() {
 		int sum = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i = 0; i < LoopIterations; i++) {
 			for (int j = 0; j < Data.Count; j++) {
 				sum += (int)Data[CollectionsHelpers.RandomIndices[j]];
 			}
@@ -47,7 +48,7 @@ public class ArrayListBenchmarks {
 		int result = 0;
 		ArrayList target = new();
 
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i = 0; i < LoopIterations; i++) {
 			for (int index = 0; index < Data.Count; index++) {
 				target.Add(index);
 			}
@@ -64,7 +65,7 @@ public class ArrayListBenchmarks {
 		int result = 0;
 		ArrayList target = new();
 
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i = 0; i < LoopIterations; i++) {
 			for (int j = 0; j < Data.Count; j++) {
 				target.Add(j);
 			}
@@ -82,7 +83,7 @@ public class ArrayListBenchmarks {
 	[Benchmark("ListCreation", "Tests allocation and initialization of an ArrayList")]
 	public static int ArrayListCreation() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i = 0; i < LoopIterations; i++) {
 			ArrayList array = new ArrayList();
 			for (int index = 0; index < Data.Count; index++) {
 				array.Add(index * 2);
@@ -97,7 +98,7 @@ public class ArrayListBenchmarks {
 	[Benchmark("ListCopy", "Tests copying an ArrayList using a foreach loop")]
 	public static int ArrayListCopyManualForeach() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i = 0; i < LoopIterations; i++) {
 			ArrayList target = new();
 			foreach (int element in Data) {
 				target.Add(element);
@@ -113,7 +114,7 @@ public class ArrayListBenchmarks {
 	[Benchmark("ListCopy", "Tests copying an ArrayList using a for loop")]
 	public static int ArrayListCopyManualFor() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i = 0; i < LoopIterations; i++) {
 			ArrayList target = new();
 			for (int index = 0; index < Data.Count; index++) {
 				target.Add(Data[index]);
@@ -129,7 +130,7 @@ public class ArrayListBenchmarks {
 	[Benchmark("ListCopy", "Tests copying an ArrayList using Clone")]
 	public static int ArrayListCopyClone() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i = 0; i < LoopIterations; i++) {
 			ArrayList target = new();
 			target = (ArrayList)Data.Clone();
 			result += target.Count;
