+++ Benchmarks/src/Collections/List/LinkedListBenchmarks.cs
@@ -8,8 +8,9 @@ namespace Benchmarks.Collections.List;
 [SuppressMessage("ReSharper", "UnusedMember.Global")]
 [SuppressMessage("ReSharper", "UnusedType.Global")]
 public class LinkedListBenchmarks {
-	public static int Iterations;
-	public static int LoopIterations;
+	public static ulong Iterations;
+	public static ulong LoopIterations;
+
 
 	public static readonly LinkedList<int> Data = new();
 
@@ -22,7 +23,7 @@ public class LinkedListBenchmarks {
 	[Benchmark("ListGet", "Tests getting values sequentially from a LinkedList (Note this uses ElementAt)")]
 	public static int LinkedListGet() {
 		int sum = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			for (int j = 0; j < Data.Count; j++) {
 				sum += Data.ElementAt(CollectionsHelpers.SequentialIndices[j]);
 			}
@@ -34,7 +35,7 @@ public class LinkedListBenchmarks {
 	[Benchmark("ListGet", "Tests getting values randomly from a LinkedList (Note this uses ElementAt)")]
 	public static int LinkedListGetRandom() {
 		int sum = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			for (int j = 0; j < Data.Count; j++) {
 				sum += Data.ElementAt(CollectionsHelpers.RandomIndices[j]);
 			}
@@ -48,7 +49,7 @@ public class LinkedListBenchmarks {
 		int result = 0;
 		LinkedList<int> target = new();
 
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			for (int index = 0; index < Data.Count; index++) {
 				target.AddLast(index);
 			}
@@ -65,7 +66,7 @@ public class LinkedListBenchmarks {
 		int result = 0;
 		LinkedList<int> target = new();
 
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			for (int index = 0; index < Data.Count; index++) {
 				target.AddFirst(index);
 			}
@@ -82,7 +83,7 @@ public class LinkedListBenchmarks {
 		int result = 0;
 		LinkedList<int> target = new();
 
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			for (int j = 0; j < Data.Count; j++) {
 				target.AddLast(j);
 			}
@@ -100,7 +101,7 @@ public class LinkedListBenchmarks {
 	[Benchmark("ListCreation", "Tests allocation and initialization of a LinkedList")]
 	public static int LinkedListCreation() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			LinkedList<int> linkedList = new LinkedList<int>();
 			for (int index = 0; index < Data.Count; index++) {
 				linkedList.AddLast(index * 2);
@@ -115,7 +116,7 @@ public class LinkedListBenchmarks {
 	[Benchmark("ListCopy", "Tests copying an LinkedList using a for loop")]
 	public static int LinkedListCopyManualFor() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			LinkedList<int> target = new();
 			LinkedListNode<int> data = Data.First;
 			for (int index = 0; index < Data.Count; index++) {
@@ -132,7 +133,7 @@ public class LinkedListBenchmarks {
 	[Benchmark("ListCopy", "Tests copying an LinkedList using a foreach loop")]
 	public static int LinkedListCopyManualForeach() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			LinkedList<int> target = new();
 			foreach (int element in Data) {
 				target.AddLast(element);
