+++ Benchmarks/src/Collections/Set/ImmutableHashSetBenchmarks.cs
@@ -7,8 +7,9 @@ namespace Benchmarks.Collections.Set;
 [SuppressMessage("ReSharper", "UnusedMember.Global")]
 [SuppressMessage("ReSharper", "UnusedType.Global")]
 public class ImmutableHashSetBenchmarks {
-	public static int Iterations;
-	public static int LoopIterations;
+	public static ulong Iterations;
+	public static ulong LoopIterations;
+
 
 	public static readonly ImmutableHashSet<int> Data;
 
@@ -20,7 +21,7 @@ public class ImmutableHashSetBenchmarks {
 	[Benchmark("SetCreation", "Tests allocation and initialization of an ImmutableHashSet")]
 	public static int ImmutableHashSetCreation() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			ImmutableHashSet<int> hashSet = ImmutableHashSet<int>.Empty;
 			for (int index = 0; index < Data.Count; index++) {
 				hashSet = hashSet.Add(index * 2);
@@ -36,7 +37,7 @@ public class ImmutableHashSetBenchmarks {
 	[Benchmark("SetGet", "Tests getting values from a ImmutableHashSet")]
 	public static int ImmutableHashSetGet() {
 		int sum = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			for (int j = 0; j < Data.Count; j++) {
 				Data.TryGetValue(j, out int value);
 				sum += value;
@@ -49,7 +50,7 @@ public class ImmutableHashSetBenchmarks {
 	[Benchmark("SetInsertion", "Tests insertion into a ImmutableHashSet")]
 	public static int ImmutableHashSetInsertion() {
 		ImmutableHashSet<int> temp = ImmutableHashSet<int>.Empty;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			temp = ImmutableHashSet<int>.Empty;
 			for (int j = 0; j < 1000; j++) {
 				temp = temp.Add(j);
@@ -62,7 +63,7 @@ public class ImmutableHashSetBenchmarks {
 	[Benchmark("SetRemoval", "Tests removal from a ImmutableHashSet")]
 	public static int ImmutableHashSetRemoval() {
 		ImmutableHashSet<int> temp = ImmutableHashSet<int>.Empty;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			for (int j = 0; j < 1000; j++) {
 				temp = temp.Add(j);
 			}
@@ -78,7 +79,7 @@ public class ImmutableHashSetBenchmarks {
 	[Benchmark("SetCopy", "Tests copying an ImmutableHashSet using a foreach loop")]
 	public static int ImmutableHashSetCopyManualForeach() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			ImmutableHashSet<int> target = ImmutableHashSet<int>.Empty;
 			foreach (int element in Data) {
 				target = target.Add(element);
@@ -94,7 +95,7 @@ public class ImmutableHashSetBenchmarks {
 	[Benchmark("SetCopy", "Tests copying an ImmutableHashSet using a for loop")]
 	public static int ImmutableHashSetCopyManualFor() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			ImmutableHashSet<int> target = ImmutableHashSet<int>.Empty;
 			for (int j = 0; j < Data.Count; j++) {
 				Data.TryGetValue(j, out int value);
