+++ Benchmarks/src/Collections/Set/ImmutableSortedSetBenchmarks.cs
@@ -7,8 +7,9 @@ namespace Benchmarks.Collections.Set;
 [SuppressMessage("ReSharper", "UnusedMember.Global")]
 [SuppressMessage("ReSharper", "UnusedType.Global")]
 public class ImmutableSortedSetBenchmarks {
-	public static int Iterations;
-	public static int LoopIterations;
+	public static ulong Iterations;
+	public static ulong LoopIterations;
+
 
 
 	public static readonly ImmutableSortedSet<int> Data;
@@ -21,7 +22,7 @@ public class ImmutableSortedSetBenchmarks {
 	[Benchmark("SetCreation", "Tests allocation and initialization of an ImmutableSortedSet")]
 	public static int ImmutableSortedSetCreation() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			ImmutableSortedSet<int> sortedSet = ImmutableSortedSet<int>.Empty;
 			for (int index = 0; index < Data.Count; index++) {
 				sortedSet = sortedSet.Add(index * 2);
@@ -37,7 +38,7 @@ public class ImmutableSortedSetBenchmarks {
 	[Benchmark("SetGet", "Tests getting values from a ImmutableSortedSet")]
 	public static int ImmutableSortedSetGet() {
 		int sum = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			for (int j = 0; j < Data.Count; j++) {
 				Data.TryGetValue(j, out int value);
 				sum += value;
@@ -50,7 +51,7 @@ public class ImmutableSortedSetBenchmarks {
 	[Benchmark("SetInsertion", "Tests insertion into a ImmutableSortedSet")]
 	public static int ImmutableSortedSetInsertion() {
 		ImmutableSortedSet<int> temp = ImmutableSortedSet<int>.Empty;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			temp = ImmutableSortedSet<int>.Empty;
 			for (int j = 0; j < 1000; j++) {
 				temp = temp.Add(j);
@@ -63,7 +64,7 @@ public class ImmutableSortedSetBenchmarks {
 	[Benchmark("SetRemoval", "Tests removal from a ImmutableSortedSet")]
 	public static int ImmutableSortedSetRemoval() {
 		ImmutableSortedSet<int> temp = ImmutableSortedSet<int>.Empty;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			for (int j = 0; j < 1000; j++) {
 				temp = temp.Add(j);
 			}
@@ -79,7 +80,7 @@ public class ImmutableSortedSetBenchmarks {
 	[Benchmark("SetCopy", "Tests copying an ImmutableSortedSet using a foreach loop")]
 	public static int ImmutableSortedSetCopyManualForeach() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			ImmutableSortedSet<int> target = ImmutableSortedSet<int>.Empty;
 			foreach (int element in Data) {
 				target = target.Add(element);
@@ -95,7 +96,7 @@ public class ImmutableSortedSetBenchmarks {
 	[Benchmark("SetCopy", "Tests copying an ImmutableSortedSet using a for loop")]
 	public static int ImmutableSortedSetCopyManualFor() {
 		int result = 0;
-		for (int i = 0; i < LoopIterations; i++) {
+		for (ulong i  = 0; i < LoopIterations; i++) {
 			ImmutableSortedSet<int> target = ImmutableSortedSet<int>.Empty;
 			for (int j = 0; j < Data.Count; j++) {
 				Data.TryGetValue(j, out int value);
