+++ Benchmarks/src/HelperObjects/Inheritance/ADoable.cs
@@ -1,5 +1,5 @@
 ﻿namespace Benchmarks.HelperObjects.Inheritance; 
 
 public abstract class ADoable {
-	public abstract int UpdateAndGetValue();
+	public abstract ulong UpdateAndGetValue();
 }
